updated rollup to work with core properly
