--- libs/core/src/another/index.spec.ts
@@ -1,7 +1,0 @@
-import { YAY } from './index';
-
-describe('Yay', () => {
-  it('should contain yay', () => {
-    expect(YAY).toContain('yay');
-  });
-});
