--- libs/core/src/another/index.ts
@@ -1,3 +1,0 @@
-import { v4 as uuid } from 'uuid';
-
-export const YAY = `yay-${uuid()}`;
