--- libs/core/src/some-file.ts
@@ -1,1 +1,0 @@
-export const somTHINGis = 'iaushbdkalsbndkas';
